EZP-24361: Extracted XMLText rendering to dedicated Twig extension
